ModuleFonbt included
This module can load a font given a png, the letters size and the string with the maping.
On the Renderer Module a function print has
been added where you can give the id from
the font, the text and the position and it
prints the text on the position using id font.